Handle predefined and builtin functions

+++ include/ast-nodes/Const.h
@@ -7,11 +7,11 @@ using namespace std;
 class Const : public Expression
 {
 public:
-Const(Type *type, int c)
+    Const(Type *type, int c)
         : Expression(type), value(c) {}
 
     string buildIR(CFG *cfg);
 
 protected:
-    int value;
+    long long value;
 };
